server/requirements.txt
# Server Dependencies for Prism DNS Server (SCRUM-12)

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP Client for PowerDNS integration
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
httpx>=0.24.0  # For FastAPI testing

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Memory monitoring for tests
psutil>=5.9.0
prometheus-client>=0.19.0

# Authentication & Security
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Email
fastapi-mail>=1.4.0
jinja2>=3.0.0
premailer>=3.10.0
html2text>=2020.1.16
aiosmtplib>=3.0.0
email-validator>=2.1.0
colorama>=0.4.6
boto3>=1.26.0
botocore>=1.29.0

# Database migrations
alembic>=1.13.0

# Rate limiting
slowapi>=0.1.9

# PostgreSQL support
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
